+++ src/components/MewOverlay/MewOverlay.vue
@@ -18,12 +18,7 @@
         class="mobile-title-block pa-3 d-flex align-center justify-space-between full-width"
       >
         <div style="min-width: 36px">
-          <v-btn
-            v-if="back"
-            icon
-            color="textBlack2"
-            @click="back"
-          >
+          <v-btn v-if="back" icon color="textBlack2" @click="back">
             <v-icon size="24">
               mdi-arrow-left
             </v-icon>
@@ -36,16 +31,8 @@
           {{ title }}
         </h3>
         <div style="min-width: 36px">
-          <v-btn
-            v-if="close"
-            icon
-            color="textBlack2"
-            @click="close"
-          >
-            <v-icon
-              size="24"
-              color="textBlack2"
-            >
+          <v-btn v-if="close" icon color="textBlack2" @click="close">
+            <v-icon size="24" color="textBlack2">
               mdi-close
             </v-icon>
           </v-btn>
@@ -67,7 +54,7 @@
         color="textBlack2"
         :class="[
           'd-flex action-btn align-center pa-3',
-          isMobile ? 'mt-3 ml-n1 mobile-btn' : 'mt-4 ml-1'
+          isMobile ? 'mt-3 ml-n1 mobile-btn' : 'mt-4 ml-1',
         ]"
         fab
         @click="back"
@@ -87,21 +74,15 @@
         :class="[
           'd-flex action-btn align-center pa-3',
           ,
-          isMobile ? 'mt-3 mr-n1 mobile-btn' : 'mt-4 mr-1'
+          isMobile ? 'mt-3 mr-n1 mobile-btn' : 'mt-4 mr-1',
         ]"
         @click="close"
       >
-        <v-icon
-          size="24"
-          color="textBlack2"
-        >
+        <v-icon size="24" color="textBlack2">
           mdi-close
         </v-icon>
       </v-btn>
-      <v-container
-        :class="['ma-0 pa-0', isMobile ? 'full-height' : '']"
-        fluid
-      >
+      <v-container :class="['ma-0 pa-0', isMobile ? 'full-height' : '']" fluid>
         <!-- ===================================================================================== -->
         <!-- White sheet (displays on the overlay - size is based on the contentSize prop) -->
         <!-- ===================================================================================== -->
@@ -125,8 +106,8 @@
                   isMobile
                     ? 'mew-heading-2 ml-3'
                     : isMobile && !back
-                      ? 'mew-heading-2 ml-4'
-                      : 'mew-subtitle'
+                    ? 'mew-heading-2 ml-4'
+                    : 'mew-subtitle',
                 ]"
               >
                 {{ title }}
@@ -139,7 +120,7 @@
             <div
               :class="[
                 'd-flex flex-column align-center justify-center',
-                isMobile ? 'px-3 pb-6 mobile-content' : 'px-8 pb-8'
+                isMobile ? 'px-3 pb-6 mobile-content' : 'px-8 pb-8',
               ]"
             >
               <slot />
@@ -150,10 +131,7 @@
         <!-- ===================================================================================== -->
         <!-- Footer -->
         <!-- ===================================================================================== -->
-        <v-row
-          justify="center"
-          class="ma-0 py-8 textMedium--text"
-        >
+        <v-row justify="center" class="ma-0 py-8 textMedium--text">
           {{ footer.text }}
           <a
             v-if="footer && footer.linkTitle && footer.link"
@@ -172,14 +150,14 @@
 
 <script>
 const sizes = {
-  small: 'small',
-  medium: 'medium',
-  large: 'large',
-  xlarge: 'xlarge'
+  small: "small",
+  medium: "medium",
+  large: "large",
+  xlarge: "xlarge",
 };
 
 export default {
-  name: 'MewOverlay',
+  name: "MewOverlay",
   props: {
     /**
      * Displays on the outside bottom of the white sheet.
@@ -190,7 +168,7 @@ export default {
       type: Object,
       default: () => {
         return {};
-      }
+      },
     },
     /**
      * Opens the overlay from
@@ -198,7 +176,7 @@ export default {
      */
     showOverlay: {
       type: Boolean,
-      default: false
+      default: false,
     },
     /**
      * Overlay title.
@@ -206,7 +184,7 @@ export default {
      */
     title: {
       type: String,
-      default: ''
+      default: "",
     },
     /**
      * Function that gets triggered
@@ -217,7 +195,7 @@ export default {
       type: Function,
       default: () => {
         return {};
-      }
+      },
     },
     /**
      * Function that gets triggered
@@ -225,7 +203,7 @@ export default {
      */
     back: {
       type: Function,
-      default: null
+      default: null,
     },
     /**
      * Applies the size of the white sheet on the overlay.
@@ -234,12 +212,12 @@ export default {
      */
     contentSize: {
       type: String,
-      default: 'small'
-    }
+      default: "small",
+    },
   },
   data() {
     return {
-      isOverlayShown: false
+      isOverlayShown: false,
     };
   },
   computed: {
@@ -250,25 +228,29 @@ export default {
       if (this.contentSize) {
         switch (this.contentSize.toLowerCase()) {
           case sizes.small:
-            return '384px';
+            return "384px";
           case sizes.medium:
-            return '504px';
+            return "504px";
           case sizes.large:
-            return '624px';
+            return "624px";
           case sizes.xlarge:
-            return '744px';
+            return "744px";
           default:
-            return '384px';
+            return "384px";
         }
       }
-      return '384px';
-    }
+      return "384px";
+    },
   },
   watch: {
     showOverlay(newVal) {
-      this.removeHtmlScrollbar();
+      if (newVal) {
+        this.removeHtmlScrollbar();
+      } else {
+        this.restoreHtmlScrollbar();
+      }
       this.isOverlayShown = newVal;
-    }
+    },
   },
   destroyed() {
     this.restoreHtmlScrollbar();
@@ -279,22 +261,22 @@ export default {
   methods: {
     // Remove html element's side bar to fix double sidebar bug
     removeHtmlScrollbar() {
-      const htmlElement = document.querySelector('html');
-      htmlElement.style.overflow = 'hidden';
+      const htmlElement = document.querySelector("html");
+      htmlElement.style.overflow = "hidden";
     },
     // Restore html element's side bar on exit
     restoreHtmlScrollbar() {
-      const htmlElement = document.querySelector('html');
+      const htmlElement = document.querySelector("html");
       htmlElement.style.overflow = null;
-    }
-  }
+    },
+  },
 };
 </script>
 
 <style lang="scss" scoped>
 .mew-overlay-container::before,
 .mew-overlay-container::after {
-  content: '';
+  content: "";
   margin: auto;
 }